+++ lib/json/ld/decoder.ex
@@ -1,5 +1,10 @@
 defmodule JSON.LD.Decoder do
   @moduledoc """
+  A decoder for JSON-LD serializations to `RDF.Dataset`s.
+
+  As for all decoders of `RDF.Serialization.Format`s, you normally won't use these
+  functions directly, but via one of the `read_` functions on the `JSON.LD` format
+  module or the generic `RDF.Serialization` module.
   """
 
   use RDF.Serialization.Decoder